Progress on dSndSourceMgr_c::createSource

+++ include/d/snd/d_snd_3d_actor.h
@@ -10,7 +10,7 @@
 
 class dSnd3DActor_c : public nw4r::snd::Sound3DActor {
 public:
-    dSnd3DActor_c(dSndSourceParam *pSourceParam, u8 sourceType);
+    dSnd3DActor_c(dSndSourceParam *pSourceParam, s32 sourceType);
 
     virtual ~dSnd3DActor_c() {}
